imdb info movies

+++ applications/admin/modules/movies/module_movies.class.php
@@ -113,38 +113,39 @@ class Module_Movies extends GW_Common_Module
 			$imdb_api['genreString'] = $imdb->getGenreString();
 			$imdb_api['description'] = $imdb->getDescription();
 			$imdb_api['plot'] = $imdb->getPlot();
-			$imdb_api['imdbID'] = $imdb->getImdbID();
+			$imdb_api['imdbID'] = (int)$imdb->getImdbID();
 			$imdb_api['poster'] = $imdb->getPoster();
 			$imdb_api['rating'] = $imdb->getRating();
 			$imdb_api['runtime'] = $imdb->getRuntime();
 			$imdb_api['title'] = $imdb->getTitle();
 			$imdb_api['year'] = $imdb->getYear();
 			
-			
-			$item->name_orig = $imdb_api['title'].' '.$imdb_api['year'];
-			$tmpfilename=GW::s('DIR/TEMP').'imdbposter_'.GW_File_Helper::cleanName($item->name_orig).'_'.date('Ymd_hmi').'.jpg';
-			
-			$data=file_get_contents($imdb_api['poster']);
-			
-			file_put_contents($tmpfilename, $data);
-			
-			$image = Array
-			    (
-			    'new_file' => $tmpfilename,
-			    'size' => filesize($tmpfilename),
-			    'original_filename' => GW_File_Helper::cleanName($item->name_orig).'.jpg',
-			);;
+			if($imdb_api['imdbID']){	
+				$item->name_orig = $imdb_api['title'].' '.$imdb_api['year'];
+				$tmpfilename=GW::s('DIR/TEMP').'imdbposter_'.GW_File_Helper::cleanName($item->name_orig).'_'.date('Ymd_hmi').'.jpg';
 
-			
-			
-			$item->set('image1', $image);	
-			$item->validate();//resizes image
-			
-			$item->update();
-			
-			unlink($tmpfilename);
-			
-			
+				$data=file_get_contents($imdb_api['poster']);
+
+				file_put_contents($tmpfilename, $data);
+
+				$image = Array
+				    (
+				    'new_file' => $tmpfilename,
+				    'size' => filesize($tmpfilename),
+				    'original_filename' => GW_File_Helper::cleanName($item->name_orig).'.jpg',
+				);;
+
+
+
+				$item->set('image1', $image);	
+				$item->validate();//resizes image
+
+				$item->update();
+
+				unlink($tmpfilename);
+			}else{
+				['error'=>'404'];
+			}
 		}else{
 			$imdb_api=['error'=>$imdb->status];
 		}
